+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Added getClientIp method to Request.
 - Added getReferrer method to Request.
 - Added getCookieValue method to Request.
+- Added setCookieValue method to Response.
 - Added ErrorControllerTrait.
 - Added isPreActionEventEnabled and isPostActionEventEnabled methods to Controller.
 - Prevented onPreActionEvent and onPreActionEvent methods from running on error controller if an exception occurred.